--- src/main/java/roomescape/controller/RoomescapeController.java
@@ -1,24 +1,0 @@
-package roomescape.controller;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class RoomescapeController {
-
-    @GetMapping("")
-    public String index() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "admin/reservation";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "admin/time";
-    }
-}
